Fixed delivery of manifest downloader

+++ sat6_scripts.spec
@@ -1,11 +1,11 @@
 Name:           sat6_scripts
-Version:        1.0.0
-Release:        1%{?dist}
+Version:        0.1
+Release:        3%{?dist}
 Summary:        Scripts to automate Satellite 6 tasks
 
 License:        GPL
 URL:            https://github.com/ggatward/sat6_scripts
-Source0:        sat6_scripts.tar.gz
+Source0:        sat6_scripts-0.1.tar.gz
 
 Requires:       python >= 2.7, PyYAML
 
@@ -24,7 +24,7 @@ Various scripts to substantially automate management tasks of Satellite 6, inclu
 
 %install
 rm -rf %{buildroot}
-mkdir -p %{buildroot}/usr/local/{bin,etc}
+mkdir -p %{buildroot}/usr/local/{bin,etc/sat6_scripts}
 mkdir -p %{buildroot}/usr/share/{doc/sat6_scripts,sat6_scripts/config}
 install -m 0644 README.md %{buildroot}/usr/share/doc/sat6_scripts/README.md
 install -m 0644 LICENSE %{buildroot}/usr/share/doc/sat6_scripts/LICENSE
@@ -39,6 +39,7 @@ install -m 0755 bin/sat_import %{buildroot}/usr/local/bin/sat_import
 install -m 0755 bin/clean_content_views %{buildroot}/usr/local/bin/clean_content_views
 install -m 0755 bin/publish_content_views %{buildroot}/usr/local/bin/publish_content_views
 install -m 0755 bin/promote_content_views %{buildroot}/usr/local/bin/promote_content_views
+install -m 0755 bin/download_manifest %{buildroot}/usr/local/bin/download_manifest
 install -m 0644 helpers.py %{buildroot}/usr/share/sat6_scripts/helpers.py
 install -m 0644 check_sync.py %{buildroot}/usr/share/sat6_scripts/check_sync.py
 install -m 0644 sat_export.py %{buildroot}/usr/share/sat6_scripts/sat_export.py
@@ -74,16 +75,24 @@ install -m 0644 download_manifest.py %{buildroot}/usr/share/sat6_scripts/downloa
 /usr/local/bin/clean_content_views
 /usr/local/bin/publish_content_views
 /usr/local/bin/promote_content_views
+/usr/local/bin/download_manifest
 
+%exclude /usr/share/sat6_scripts/*.pyc
+%exclude /usr/share/sat6_scripts/*.pyo
 
 %post
-ln -s /usr/share/sat6_scripts/config/config.yml /usr/local/etc/sat6_scripts/config.yml
-ln -s /usr/share/sat6_scripts/config/exports.yml /usr/local/etc/sat6_scripts/exports.yml
-
+# Only run on initial install
+if [ $1 -eq 1 ]; then
+  ln -s /usr/share/sat6_scripts/config/config.yml /usr/local/etc/sat6_scripts/config.yml
+  ln -s /usr/share/sat6_scripts/config/exports.yml /usr/local/etc/sat6_scripts/exports.yml
+fi
 
 %postun
-unlink /usr/local/etc/sat6_scripts/config.yml
-unlink /usr/local/etc/sat6_scripts/exports.yml
+# Only run on a complete uninstall
+if [ $1 -eq 0 ]; then
+  unlink /usr/local/etc/sat6_scripts/config.yml
+  unlink /usr/local/etc/sat6_scripts/exports.yml
+fi
 
 
 %changelog
